requirements.txt
streamlit
pandas
plotly
smartapi-python
pyotp
websocket-client
pytz
logzero
requests